+++ src/cfg.h
@@ -75,9 +75,6 @@ enum MY_IDS
 
     ID_MENU_PRINTPAGE,
     ID_MENU_PRINTFILE,
-    ID_MENU_MYLOG_SAVE,
-    ID_MENU_MYLOG_CLEAR,
-    ID_MENU_MYLOG_HIDE,
     ID_MENU_OLD_TO_DBASE,
     ID_MENU_DBASE_TO_OLD,
     ID_MENU_OLD_DBASE,      // 'original' way of saving game-data: game.* files